server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["sam-ayo <sam-ayo>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
beanie = "^1.26.0"
python-dotenv = "^1.0.1"
firebase-admin = "^6.5.0"
pydantic-settings = "^2.2.1"
authlib = "^1.3.0"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
